docker/util.sh: add --no-rsync-daemon
Enable building without the rsync daemon.

This is preparation for being able
to use ccache and Docker caching
without syncing with the home directory.
As a bonus, no rsync daemon means
the build can use the default network
instead of host, so more encapsulation.

+++ docker/util.sh
@@ -17,7 +17,7 @@ fi
 
 usage()
 {
-    echo "Usage: build.sh [--push] [--docker-repository repo] [--tag name] [--gdal tag|sha1|master] [--gdal-repository repo] [--proj tag|sha1|master] [--release] [--docker-cache|--no-docker-cache]"
+    echo "Usage: build.sh [--push] [--docker-repository repo] [--tag name] [--gdal tag|sha1|master] [--gdal-repository repo] [--proj tag|sha1|master] [--release] [--docker-cache|--no-docker-cache] [--no-rsync-daemon]"
     # Non-documented: --test-python
     echo ""
     echo "--push: push image to docker repository (defaults to ghcr.io)"
@@ -27,6 +27,7 @@ usage()
     echo "--gdal-repository repo: github repository. Defaults to OSGeo/gdal"
     echo "--release: Whether this is a release build. In which case --gdal tag must be used."
     echo "--docker-cache/--no-docker-cache: instruct Docker to build with/without using its cache. Defaults to no cache for release builds."
+    echo "--no-rsync-daemon: do not use the rsync daemon to save build cache in home directory."
     echo "--with-debug-symbols/--without-debug-symbols. Whether to include debug symbols. Only applies to ubuntu-full, default is to include for non-release builds."
     exit 1
 }
@@ -107,6 +108,11 @@ do
         ;;
         --no-docker-cache)
             DOCKER_CACHE_PARAM="--no-cache"
+            shift
+        ;;
+
+        --no-rsync-daemon)
+            NO_RSYNC_DAEMON=1
             shift
         ;;
 
@@ -250,6 +256,7 @@ if test "${RELEASE}" = "yes"; then
     fi
 
 else
+    BUILD_ARGS+=("--build-arg" "BUILDKIT_INLINE_CACHE=1")
 
     IMAGE_NAME_WITH_ARCH="${REPO_IMAGE_NAME}"
     if test "${IMAGE_NAME}" = "osgeo/gdal:ubuntu-full-latest" \
@@ -268,21 +275,23 @@ else
         GDAL_RELEASE_DATE=$(date "+%Y%m%d")
     fi
     echo "Using GDAL_RELEASE_DATE=${GDAL_RELEASE_DATE}"
+    BUILD_ARGS+=("--build-arg" "GDAL_RELEASE_DATE=${GDAL_RELEASE_DATE}")
 
-    RSYNC_DAEMON_CONTAINER=gdal_rsync_daemon
-    HOST_CACHE_DIR="$HOME/gdal-docker-cache"
+    if [ ! -v NO_RSYNC_DAEMON ]; then
+        RSYNC_DAEMON_CONTAINER=gdal_rsync_daemon
+        HOST_CACHE_DIR="$HOME/gdal-docker-cache"
 
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/proj/x86_64"
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/proj/aarch64"
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/gdal/x86_64"
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/gdal/aarch64"
-    mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/spatialite"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/proj/x86_64"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/proj/aarch64"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/gdal/x86_64"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/gdal/aarch64"
+        mkdir -p "${HOST_CACHE_DIR}/${TARGET_IMAGE}/spatialite"
 
-    # Start a Docker container that has a rsync daemon, mounting HOST_CACHE_DIR
-    if ! docker ps | grep "${RSYNC_DAEMON_CONTAINER}"; then
-        RSYNC_DAEMON_IMAGE=osgeo/gdal:gdal_rsync_daemon
-        docker rmi "${RSYNC_DAEMON_IMAGE}" 2>/dev/null || true
-        docker $(build_cmd) -t "${RSYNC_DAEMON_IMAGE}" - <<EOF
+        # Start a Docker container that has a rsync daemon, mounting HOST_CACHE_DIR
+        if ! docker ps | grep "${RSYNC_DAEMON_CONTAINER}"; then
+            RSYNC_DAEMON_IMAGE=osgeo/gdal:gdal_rsync_daemon
+            docker rmi "${RSYNC_DAEMON_IMAGE}" 2>/dev/null || true
+            docker $(build_cmd) -t "${RSYNC_DAEMON_IMAGE}" - <<EOF
 FROM alpine
 
 VOLUME /opt/gdal-docker-cache
@@ -300,24 +309,21 @@ CMD rsync --daemon --port 23985 && while sleep 1; do true; done
 
 EOF
 
-        THE_UID=$(id -u "${USER}")
-        THE_GID=$(id -g "${USER}")
+            THE_UID=$(id -u "${USER}")
+            THE_GID=$(id -g "${USER}")
 
-        docker run -d -u "${THE_UID}:${THE_GID}" --rm \
-            -v "${HOST_CACHE_DIR}":/opt/gdal-docker-cache \
-            --name "${RSYNC_DAEMON_CONTAINER}" \
-            --network host \
-            "${RSYNC_DAEMON_IMAGE}"
-
+            docker run -d -u "${THE_UID}:${THE_GID}" --rm \
+                   -v "${HOST_CACHE_DIR}":/opt/gdal-docker-cache \
+                   --name "${RSYNC_DAEMON_CONTAINER}" \
+                   --network host \
+                   "${RSYNC_DAEMON_IMAGE}"
+        fi
+        RSYNC_REMOTE="rsync://127.0.0.1:23985/gdal-docker-cache/${TARGET_IMAGE}"
+        BUILD_ARGS+=(
+            "--build-arg" "RSYNC_REMOTE=${RSYNC_REMOTE}" \
+            "--network" "host" \
+        )
     fi
-
-    RSYNC_REMOTE="rsync://127.0.0.1:23985/gdal-docker-cache/${TARGET_IMAGE}"
-
-    BUILD_ARGS+=(
-        "--build-arg" "GDAL_RELEASE_DATE=${GDAL_RELEASE_DATE}" \
-        "--build-arg" "RSYNC_REMOTE=${RSYNC_REMOTE}" \
-        "--build-arg" "BUILDKIT_INLINE_CACHE=1" \
-    )
 
     if test "${BASE_IMAGE}" != ""; then
         BUILD_ARGS+=("--build-arg" "BASE_IMAGE=${BASE_IMAGE}")
@@ -358,7 +364,7 @@ EOF
       fi
     fi
 
-    docker $(build_cmd) --network=host "${BUILD_ARGS[@]}" -t "${IMAGE_NAME_WITH_ARCH}" "${SCRIPT_DIR}"
+    docker $(build_cmd) "${BUILD_ARGS[@]}" -t "${IMAGE_NAME_WITH_ARCH}" "${SCRIPT_DIR}"
 
     if test "${DOCKER_BUILDX}" != "buildx"; then
         check_image "${IMAGE_NAME_WITH_ARCH}"
